feat: move to viper config management, add service collection, refactor code structure

+++ internal/bot/workers/fasthttp_server.go
@@ -20,7 +20,7 @@ func gracefulShutdownServer(srv *fasthttp.Server) {
 	}
 }
 
-func StartServer(ctx context.Context, srv *fasthttp.Server, port int16) {
+func StartServer(ctx context.Context, srv *fasthttp.Server, port int) {
 	defer gracefulShutdownServer(srv)
 
 	go func() {
